+++ CHANGELOG.md
@@ -9,6 +9,7 @@ LUYA CHANGELOG
 - `#271`: removed twig function `linkActive()` added global variable `activeUrl`.
 - `#244`: removed NgRest `required()` plugin.
 - `#215`: removed `luya\base\PageController` and replace with basic `luya\base\Controller`. `$this->context->pageTitle` is now invalid.
+- `#221`: removed `presql` command and replaced with `migrate` command.
 
 1.0.0-alpha10 (9. Jul 2015)
 ---------------------------